+++ public/ide/modules/ide/js/parts/stage/stageController.js
@@ -236,7 +236,6 @@
 
             if (oldOperate){
                 var nowOperate=ProjectService.SaveCurrentOperate();
-
                 var operates={
                     undoOperate:oldOperate,
                     redoOperate:nowOperate
@@ -273,6 +272,7 @@
 
             var pageIndex=-1;
 
+            //console.log('pages',$scope.project.pages);
             _.forEach($scope.project.pages,function(page,index){
                 if (page.id==ProjectService.getCurrentPage().id){
                     pageIndex=index;
@@ -492,11 +492,11 @@
 
 
             ProjectService.OnLayerDoubleClicked(_target.id,function () {
-                // $timeout(function () {
-                //     layerDoubleClicking=false;
-                //
-                // },100);
-                layerDoubleClicking=false;
+                $timeout(function () {
+                    layerDoubleClicking=false;
+                
+                },0);
+                // layerDoubleClicking=false;
                 //更新缩略图
                 $scope.$emit('ChangeCurrentPage');
 
@@ -562,9 +562,8 @@
             }
         }
         function releaseLayer(event){
-
             if (layerDoubleClicking){
-                console.log('双击中');
+                console.log('双击layer中');
                 return;
             }
             var isDbClick=false;    //这次点击是否是双击