+++ routes/routeBlog.js
@@ -2,6 +2,7 @@
  * Created by changecheng on 2017/2/9.
  */
 var BlogModel = require('../db/models/BlogModel');
+var UserModel = require('../db/models/UserModel');
 var CommentModel = require('../db/models/CommentModel');
 var path = require('path');
 var fs = require('fs');
@@ -423,13 +424,30 @@ BlogRoute.getBlogData = function (req, res) {
                                 var info = {};
                                 info.authorId = _comment.authorId;
                                 info.content = _comment.content;
-                                info.createTime = _comment.createTime;
+                                info.createTime = _comment.createTime.toLocaleString();
                                 info.authorName = _comment.authorName;
                                 info._id = _comment._id;
                                 return info;
                             });
                             result.comments = comments;
-                            res.end(JSON.stringify(result))
+                            var _user = req.session.user;
+                            if(_user&&_user.username&&_user.id){
+                                UserModel.findById(_user.id,function(err,data){
+                                    if(err){
+                                        console.log('err',err);
+                                        result.userInfo = null;
+                                    }else{
+                                        result.userInfo = {
+                                            id:data.id,
+                                            type:data.type
+                                        }
+                                    }
+                                    res.end(JSON.stringify(result));
+                                })
+                            }else{
+                                result.userInfo = null;
+                                res.end(JSON.stringify(result))
+                            }
                             //res.send(result);
                         }
                     });
@@ -620,8 +638,7 @@ BlogRoute.postComment = function(req,res){
         })
         //req.send(data);
     }else{
-        //res.send('not login');
-        res.redirect('/user/login')
+        res.send('not login')
     }
 
 };